Updating licensing info and add a test.
