+++ .gitignore
@@ -3,3 +3,10 @@
 .gradle
 /build/
 !gradle/wrapper/gradle-wrapper.jar
+
+### IntelliJ IDEA ###
+.idea
+*.iws
+*.iml
+*.ipr
+/out
