+++ LICENSE
@@ -1,4 +1,4 @@
-  Copyright 2017 Smarter Balanced Licensed under the
+  Copyright 2018 Smarter Balanced Licensed under the
 	Educational Community License, Version 2.0 (the "License"); you may
 	not use this file except in compliance with the License. You may
 	obtain a copy of the License at
