+++ src/main/java/org/opentestsystem/ap/imrt/iss/IssApplication.java
@@ -1,0 +1,11 @@
+package org.opentestsystem.ap.imrt.iss;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class IssApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(IssApplication.class, args);
+    }
+}
