+++ src/main/java/org/opentestsystem/ap/imrt/iss/controller/HelloController.java
@@ -1,0 +1,14 @@
+package org.opentestsystem.ap.imrt.iss.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HelloController {
+
+    @RequestMapping("/")
+    public String index() {
+        return("IMRT Item Ingest Service");
+    }
+
+}
